// types/receives.ts
export type Receives = {
  birthPlace?: string;
  certification?: string;
  comments?: string;
  dobDD?: string;
  dobMM?: string;
  dobYY?: string;
  dream?: string;
  experiences?: {
    experience: string;
    years: string;
  }[];
  freeImageUrl0?: string;
  freeImageUrl1?: string;
  freeImageUrl2?: string;
  hobby?: string;
  homeAddress?: string;
  jobTitle?: string;
  language?: string;
  profileImageUrl?: string;
  resumes?: {
    companyName: string;
    employmentStatus: string;
    workEnd: string;
    workStart: string;
  }[];
  school?: string;
  scout?: string;
  strongArea?: string;
  subjectArea?: string;
  userName?: string;
  id:string;
}